Added files from project
This project was backed up via generating backups from the make backup target. Now that it is complete, it is simply being uploaded to GitHub to demonstrate the code as an example. Note that this was specifically designed to run on the provided Linux server from UTD; the first order of business to make this program platform-independent would be to update the Makefile with a more portable solution.